requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosmtpd==1.4.2
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
atpublic==3.0.1
attrs==21.4.0
Babel==2.9.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.17.1
esbonio==0.10.3
filelock==3.5.0
frozenlist==1.3.0
geoip2==4.5.0
h11==0.13.0
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
MarkupSafe==2.0.1
maxminddb==2.2.0
multidict==5.2.0
numpy==1.22.1
outcome==1.1.0
packaging==21.3
platformdirs==2.5.0
pycparser==2.21
pydantic==1.8.2
pygls==0.11.3
Pygments==2.11.2
PyJWT==2.4.0
pymemcache==3.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyspellchecker==0.6.3
pytz==2021.3
PyYAML==6.0
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tblib==1.7.0
trio==0.19.0
trio-websocket==0.9.2
typeguard==2.13.3
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.1
whitenoise==6.1.0
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
